feat: First service to add/get/delete bookmarks
There wasn't a database to store bookmarks.

PostgreSQL has been selected for the database. It required to configure the integration tests
to use it.

A service has been created and tested. The integration test of this service is a scenario
similar to what a developer would do to test it by hand: clear everything, create bookmarks and
check what was created.

For the moment, the service doesn't do more than calling the repository.